+++ Libplanet.Net/Consensus/Context.cs
@@ -192,7 +192,10 @@ namespace Libplanet.Net.Consensus
 
             _contextTimeoutOption = contextTimeoutOptions ?? new ContextTimeoutOption();
 
-            _logger.Debug("Created Context for height #{Height}, round #{Round}", Height, Round);
+            _logger.Information(
+                "Created Context for height #{Height}, round #{Round}",
+                Height,
+                Round);
         }
 
         /// <summary>
@@ -367,71 +370,7 @@ namespace Libplanet.Net.Consensus
             {
                 Exception exception = _blockChain.ValidateBlock(block);
                 bool isValid = exception is null;
-
-                if (!isValid)
-                {
-                    _logger.Debug(
-                        exception,
-                        "BlockHeader #{Index} {BlockHash} is invalid",
-                        block.Index,
-                        block.Hash);
-
-                    _blockHashCache.AddReplace(block.Hash, false);
-                    return false;
-                }
-
-                try
-                {
-                    // Skipping the action evaluations if itself is a proposer.
-                    if (!IsCurrentRoundProposer())
-                    {
-                        // Neeed to get txs from store, lock?
-                        // TODO: Remove ChainId, enhancing lock management.
-                        // changing exception be returning to thrown.
-                        _blockChain._rwlock.EnterUpgradeableReadLock();
-
-                        try
-                        {
-                            _blockChain.ValidateBlockNonces(
-                                block.Transactions
-                                    .Select(tx => tx.Signer)
-                                    .Distinct()
-                                    .ToDictionary(
-                                        signer => signer,
-                                        signer => _blockChain.Store.GetTxNonce(
-                                            _blockChain.Id, signer)),
-                                block);
-
-                            if (_blockChain.Policy.ValidateNextBlock(
-                                _blockChain, block) is { } bpve)
-                            {
-                                throw bpve;
-                            }
-
-                            foreach (var tx in block.Transactions)
-                            {
-                                if (_blockChain.Policy.ValidateNextBlockTx(
-                                    _blockChain, tx) is { } txve)
-                                {
-                                    throw txve;
-                                }
-                            }
-
-                            _blockChain.ValidateBlockStateRootHash(block, out _);
-                        }
-                        finally
-                        {
-                            _blockChain._rwlock.ExitUpgradeableReadLock();
-                        }
-                    }
-                }
-                catch (Exception e)
-                {
-                    exception = e;
-                }
-
-                isValid = exception is null;
-                _logger.Debug(
+                _logger.Information(
                     exception,
                     "Block #{Index} {Hash} is valid? {Bool}",
                     block.Index,