+++ CHANGES.md
@@ -4,7 +4,15 @@ Libplanet changelog
 Version 0.53.4
 --------------
 
-To be released.
+Released on April 4, 2023.
+
+ -  Fixed version resolution of typescript libraries for dependencies located
+    in the monorepo.  [[#3050], [#3052]]
+ -  Exported the omitted `signTx()` and `encodeSignedTx()` functions in
+    *@planetarium/tx*.  [[#3052]]
+
+[#3050]: https://github.com/planetarium/libplanet/issues/3050
+[#3052]: https://github.com/planetarium/libplanet/pull/3052
 
 
 Version 0.53.3